// Year 2/Semester 1/Web Development/Proiect/UI/blog/src/app/features/blog-post/models/blogpost.model.ts
import { Category } from "../../category/models/category.model";

export interface BlogPost{
    title: string;
    shortDescription: string;
    content: string;
    featuredImageUrl: string;
    urlHandle: string;
    author: string;
    publishedDate: Date;
    isVisible: boolean;
    categories: Category[];
}